requirements.txt
pandas
pip
streamlit
plotly
python-dotenv
boto3
altair
fsspec
